Add Hopkins SOCS forward model and baseline reference numbers
- New differentiable Hopkins partial-coherence imaging model
  (src/openlithohub/_utils/hopkins.py) with SVD-truncated SOCS,
  circular/annular/dipole/quasar illumination, defocus support,
  and per-(params, grid) kernel caching.
- LevelSetILTModel.forward_model='hopkins' opt-in switch with
  HopkinsParams override; default remains 'gaussian' for
  back-compat.
- differentiable_threshold helper exported from
  openlithohub._utils.
- scripts/generate_baselines.py runs dummy-identity, levelset-ilt,
  and neural-ilt across eight synthetic 64x64 layouts (or
  LithoBench when --data-root is given) and writes
  baselines/results.{json,md}.
- README, docs/benchmarks.md, docs/api/models.md, mkdocs.yml,
  and CHANGELOG updated to reflect both additions.
- Fix pre-existing mypy [no-any-return] in
  benchmark/compliance/mrc.py:166.

Tests: 247 pass (17 new for hopkins, 3 new for levelset hopkins
mode); ruff/format/mypy clean.

+++ src/openlithohub/benchmark/compliance/mrc.py
@@ -163,7 +163,7 @@ def _menger_curvature(p0: np.ndarray, p1: np.ndarray, p2: np.ndarray) -> float:
     area2 = abs(cross)
     if area2 < 1e-12:
         return 0.0
-    return 2.0 * area2 / (a * b * c)
+    return float(2.0 * area2 / (a * b * c))
 
 
 def _smooth_loop(loop: np.ndarray, window: int) -> np.ndarray:
